rate limiter can support float sec

+++ src/rtbnext/core/rate_limiter.py
@@ -21,7 +21,7 @@ RateLimiterMode = Literal[ "burst", "spread" ]
 class RateLimiter:
     """Async rate limiter supporting burst and spread strategies."""
 
-    def __init__( self, max_requests: int, per_seconds: int ) -> None:
+    def __init__( self, max_requests: int, per_seconds: float ) -> None:
         self._max_requests = max_requests
         self._per_seconds = per_seconds
         self._interval = self._per_seconds / max_requests
